+++ Kernel/CMakeLists.txt
@@ -32,7 +32,7 @@ set(KERNEL_SOURCES
     Devices/RandomDevice.cpp
     Devices/SB16.cpp
     Devices/SerialDevice.cpp
-    Devices/UHCIController.cpp
+    Devices/USB/UHCIController.cpp
     Devices/VMWareBackdoor.cpp
     Devices/ZeroDevice.cpp
     Storage/Partition/DiskPartition.cpp
